requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
pytest
# MCP (Model Context Protocol) support
mcp>=0.1.0